feat: split field_splitting into two files, add doxygen comments
